refactor: async context manager
