--- packages/async-hooks-context-manager/CHANGELOG.md
@@ -1,88 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [3.7.0](https://github.com/midwayjs/midway/compare/v3.6.1...v3.7.0) (2022-10-29)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-# [3.6.0](https://github.com/midwayjs/midway/compare/v3.5.3...v3.6.0) (2022-10-10)
-
-### Features
-
-- add guard ([#2345](https://github.com/midwayjs/midway/issues/2345)) ([1b952a1](https://github.com/midwayjs/midway/commit/1b952a1b09adbb88ff3cff9a2974eb1e37ce53a5))
-
-## [3.5.3](https://github.com/midwayjs/midway/compare/v3.5.2...v3.5.3) (2022-09-25)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-## [3.5.1](https://github.com/midwayjs/midway/compare/v3.5.0...v3.5.1) (2022-09-06)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-# [3.5.0](https://github.com/midwayjs/midway/compare/v3.4.13...v3.5.0) (2022-08-29)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-## [3.4.13](https://github.com/midwayjs/midway/compare/v3.4.12...v3.4.13) (2022-08-24)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-## [3.4.12](https://github.com/midwayjs/midway/compare/v3.4.11...v3.4.12) (2022-08-20)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-## [3.4.11](https://github.com/midwayjs/midway/compare/v3.4.10...v3.4.11) (2022-08-16)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-## [3.4.10](https://github.com/midwayjs/midway/compare/v3.4.9...v3.4.10) (2022-08-12)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-## [3.4.9](https://github.com/midwayjs/midway/compare/v3.4.8...v3.4.9) (2022-08-10)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-## [3.4.7](https://github.com/midwayjs/midway/compare/v3.4.6...v3.4.7) (2022-08-01)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-## [3.4.6](https://github.com/midwayjs/midway/compare/v3.4.5...v3.4.6) (2022-07-31)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-## [3.4.4](https://github.com/midwayjs/midway/compare/v3.4.3...v3.4.4) (2022-07-25)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-## [3.4.3](https://github.com/midwayjs/midway/compare/v3.4.2...v3.4.3) (2022-07-21)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-## [3.4.1](https://github.com/midwayjs/midway/compare/v3.4.0...v3.4.1) (2022-07-20)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-# [3.4.0](https://github.com/midwayjs/midway/compare/v3.4.0-beta.12...v3.4.0) (2022-07-20)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-# [3.4.0-beta.12](https://github.com/midwayjs/midway/compare/v3.4.0-beta.11...v3.4.0-beta.12) (2022-07-20)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-# [3.4.0-beta.11](https://github.com/midwayjs/midway/compare/v3.4.0-beta.10...v3.4.0-beta.11) (2022-07-19)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-# [3.4.0-beta.10](https://github.com/midwayjs/midway/compare/v3.4.0-beta.9...v3.4.0-beta.10) (2022-07-18)
-
-**Note:** Version bump only for package @midwayjs/async-hooks-context-manager
-
-# [3.4.0-beta.9](https://github.com/midwayjs/midway/compare/v3.4.0-beta.8...v3.4.0-beta.9) (2022-07-14)
-
-### Features
-
-- add context manager with middleware ([#2116](https://github.com/midwayjs/midway/issues/2116)) ([99ba506](https://github.com/midwayjs/midway/commit/99ba506b82b1061af26bf333892ae90b654a7b31))
