--- packages/async-hooks-context-manager/README.md
@@ -1,9 +1,0 @@
-# midway-async-hooks-context-manager
-
-this is a sub package for midway.
-
-Document: [https://midwayjs.org](https://midwayjs.org)
-
-## License
-
-[MIT]((http://github.com/midwayjs/midway/blob/master/LICENSE))
