--- packages/async-hooks-context-manager/jest.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  testPathIgnorePatterns: ['<rootDir>/test/fixtures'],
-  coveragePathIgnorePatterns: ['<rootDir>/test/', '<rootDir>/dist/'],
-  setupFilesAfterEnv: ['./jest.setup.js'],
-  coverageProvider: 'v8',
-};
