--- packages/async-hooks-context-manager/jest.setup.js
@@ -1,2 +1,0 @@
-process.env.MIDWAY_TS_MODE = 'true';
-jest.setTimeout(30000);
