--- packages/async-hooks-context-manager/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@midwayjs/async-hooks-context-manager",
-  "version": "3.18.0",
-  "description": "midway async hooks context manager",
-  "main": "dist/index.js",
-  "typings": "dist/index.d.ts",
-  "scripts": {
-    "build": "tsc",
-    "test": "node -r ts-node/register ../../node_modules/jest/bin/jest.js --runInBand",
-    "cov": "node -r ts-node/register ../../node_modules/jest/bin/jest.js --runInBand --coverage --forceExit"
-  },
-  "keywords": [
-    "midway",
-    "async",
-    "context manager"
-  ],
-  "files": [
-    "dist/**/*.js",
-    "dist/**/*.d.ts"
-  ],
-  "license": "MIT",
-  "devDependencies": {
-    "@midwayjs/core": "workspace:^"
-  },
-  "author": "Harry Chen <czy88840616@gmail.com>",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/midwayjs/midway.git"
-  },
-  "engines": {
-    "node": ">=12.17.0"
-  },
-  "gitHead": "a603d2348d6141f8f723901498f03a162a037708"
-}
