--- packages/async-hooks-context-manager/src/index.ts
@@ -1,3 +1,0 @@
-export * from './asyncHooksContextManager';
-export * from './asyncLocalStorageContextManager';
-export { isSemverGreaterThanOrEqualTo, createContextManager } from './util';
