--- packages/async-hooks-context-manager/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compileOnSave": true,
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist"
-  },
-  "include": [
-    "./src/**/*.ts"
-  ]
-}
