+++ packages/core/package.json
@@ -31,7 +31,6 @@
   },
   "dependencies": {
     "@midwayjs/glob": "^1.0.2",
-    "class-transformer": "0.5.1",
     "picomatch": "2.3.1",
     "reflect-metadata": "0.2.2"
   },
