+++ packages/mock/src/creator.ts
@@ -41,7 +41,6 @@ import * as http from 'http';
 import * as https from 'https';
 import * as yaml from 'js-yaml';
 import * as getRawBody from 'raw-body';
-import { createContextManager } from '@midwayjs/async-hooks-context-manager';
 
 const debug = debuglog('midway:debug');
 
@@ -225,7 +224,6 @@ export async function create<
     await initializeGlobalApplicationContext({
       ...options,
       appDir,
-      asyncContextManager: createContextManager(),
       loggerFactory: loggers,
     });
